Code refactor with AppTabNav.
