+++ js/src/components/app-tab-nav/index.scss
@@ -1,10 +1,10 @@
-.gla-tab-nav__tabs {
+.app-tab-nav__tabs {
 	display: flex;
 	align-items: stretch;
 	box-shadow: inset 0 -1px 0 #ccc;
 	margin-bottom: var(--main-gap);
 
-	.gla-tab-nav__tabs-item {
+	&-item {
 		background: transparent;
 		border: none;
 		box-shadow: none;
